Update animations
Updated the configuration of the large and small turbine masts so that
their animations work properly in all sections.

Updated the lock variables so that they are true while the tweening
animations are active.  This lock variable is used to prevent the user
from submitting multiple calls to activate the animations before the
previous call has finished.

Added tweening camera movements between sections.  Pressing the number
keys 0 through 4 will activate the appropriate section and rotate the
camera into the default position over that section.  For convenience 0
and 4 reference the same section.